Solution for Day8. (#15)
Solution for Day8.

In the first part a network of nodes is traversed from a start node to an end node with a given path. The steps to reach the final node are counted and this number gives the solution.

The second part complicates this significantly by requiring to start from six nodes using the same path at the same time where each of those node needs a different amount of steps to reach an end node. Thus, getting to a state where all nodes are end nodes requires many cycles of the same path and trillions of steps. This means any iterative/brute force attempt to solve the puzzle takes forever and a more straightforward way to solve the problem is to first calculate the amount of steps for each start node and then find the LCM (Least Common Multiple).

General Overview:

- "aoc8.py" is the entry point that prints the solutions for part one and two.
- "Network" class contains all the required methods to solve the puzzle (including technically correct but super inefficient iterative brute force methods)
- "Counter" is a new utility helper to pass a mutable counter
- "benchmark_helper" provides benchmark printouts for function runtime